refactor(api): stand up handlers shim + shared + cache-invalidation + site domain (#35)
